Update.
2000-01-05  Andreas Schwab  <schwab@suse.de>

	* iconv/iconv_prog.c (process_block): Add cast to avoid warning.

	* locale/programs/linereader.c: Remove #include of stringtrans.h.

+++ iconv/iconv_prog.c
@@ -372,7 +372,7 @@ conversion stopped due to problem in writing the output"));
 	    {
 	    case EILSEQ:
 	      error (0, 0, _("illegal input sequence at position %ld"),
-		     addr - start);
+		     (long) (addr - start));
 	      break;
 	    case EINVAL:
 	      error (0, 0, _("\
